Crate pallet_dapps_staking

dApps Staking Module

The dApps staking module manages era, total amounts of rewards and how to distribute.

Re-exports

pub use traits::*;

pub use pallet::pallet::*;

pub use weights::WeightInfo;

Modules

benchmarking

pallet

Dapps staking FRAME Pallet.

traits

weights

Autogenerated weights for pallet_dapps_staking

Structs

EraRewardAndStake

A record for total rewards and total amount staked for an era

EraStakingPoints

Used to split total EraPayout among contracts. Each tuple (contract, era) has this structure. This will be used to reward contracts developer and his stakers.

Enums

Forcing

Mode of era-forcing.

Type Definitions

BalanceOf

EraIndex

Counter for the number of eras that have passed.

SessionIndex

Simple index type with which we can count sessions.