Enum pallet_dapps_staking::Error

pub enum Error<T: Trait> {
    NotController,
    NotStash,
    AlreadyBonded,
    AlreadyPaired,
    EmptyTargets,
    DuplicateIndex,
    InvalidSlashIndex,
    InsufficientValue,
    NoMoreChunks,
    NoUnlockChunk,
    FundedTarget,
    InvalidEraToReward,
    InvalidNumberOfNominations,
    NotSortedAndUnique,
    EmptyNominateTargets,
    NotOperatedContracts,
    NotEnoughStaking,
    // some variants omitted
}

Error for the staking module.

Variants

NotController

Not a controller account.

NotStash

Not a stash account.

AlreadyBonded

Stash is already bonded.

AlreadyPaired

Controller is already paired.

EmptyTargets

Targets cannot be empty.

DuplicateIndex

Duplicate index.

InvalidSlashIndex

Slash record index out of bounds.

InsufficientValue

Can not bond with value less than minimum balance.

NoMoreChunks

Can not schedule more unlock chunks.

NoUnlockChunk

Can not rebond without unlocking chunks.

FundedTarget

Attempting to target a stash that still has funds.

InvalidEraToReward

Invalid era to reward.

InvalidNumberOfNominations

Invalid number of nominations.

NotSortedAndUnique

Items are not sorted and unique.

EmptyNominateTargets

Targets must be latest 1.

NotOperatedContracts

Targets must be operated contracts

NotEnoughStaking

The nominations amount more than active staking amount.