Function pallet_dapps_staking::pallet::pallet::dispatchables::bond_and_stake

pub fn bond_and_stake<T: Config>(contract_id: T::SmartContract, value: Balance)

Lock up and stake balance of the origin account.

value must be more than the minimum_balance specified by MinimumStakingAmount unless account already has bonded value equal or more than ‘minimum_balance’.

The dispatch origin for this call must be Signed by the staker’s account.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::bond_and_stake.