Trait pallet_dapps_staking::pallet::pallet::Config

pub trait Config: Config {

    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber> + ReservableCurrency<Self::AccountId>;
    type SmartContract: IsContract + Parameter + Member;
    type BlockPerEra: Get<BlockNumberFor<Self>>;
    type RegisterDeposit: Get<BalanceOf<Self>>;
    type DeveloperRewardPercentage: Get<Perbill>;
    type MaxNumberOfStakersPerContract: Get<u32>;
    type MinimumStakingAmount: Get<BalanceOf<Self>>;
    type HistoryDepth: Get<u32>;
    type BonusEraDuration: Get<u32>;
    type PalletId: Get<PalletId>;
    type MinimumRemainingAmount: Get<BalanceOf<Self>>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber> + ReservableCurrency<Self::AccountId>

The staking balance.

type SmartContract: IsContract + Parameter + Member

type BlockPerEra: Get<BlockNumberFor<Self>>

Number of blocks per era.

type RegisterDeposit: Get<BalanceOf<Self>>

Minimum bonded deposit for new contract registration.

type DeveloperRewardPercentage: Get<Perbill>

Percentage of reward paid to developer.

type MaxNumberOfStakersPerContract: Get<u32>

Maximum number of unique stakers per contract.

type MinimumStakingAmount: Get<BalanceOf<Self>>

Minimum amount user must stake on contract. User can stake less if they already have the minimum staking amount staked on that particular contract.

type HistoryDepth: Get<u32>

Number of eras that are valid when claiming rewards.

All the rest will be either claimed by the treasury or discarded.

type BonusEraDuration: Get<u32>

Number of eras of doubled claim rewards.

type PalletId: Get<PalletId>

Dapps staking pallet Id

type MinimumRemainingAmount: Get<BalanceOf<Self>>

Minimum amount that should be left on staker account after staking.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.