Struct pallet_dapps_staking::ErasStakingPoints

pub struct ErasStakingPoints<T: Trait>(_);

Rewards of stakers for contracts(called by "Dapps Nominator") at era.

This is keyed first by the era index, 2nd keyed contract account to allow the stash account. Rewards for the last HISTORY_DEPTH eras.

If reward hasn't been set or has been removed then 0 reward is returned.