Crate pallet_dapps_staking

Plasm Staking Module

The Plasm staking module manages era, total amounts of rewards and how to distribute.

Re-exports

pub use parameters::StakingParameters;

pub use rewards::ComputeRewardsForDapps;

Modules

parameters
Plasm Staking Parameters

rewards
This is explains plasm inflation models. The staking has 2 kinds.

Structs

Bonded
Map from all locked "stash" accounts to the controller account.

ContractsUntreatedEra

EraStakingPoints
Reward points of an era. Used to split era total payout between dapps rewards.

ErasContractsParameters
Similarly to ErasStakers this holds the parameters of contracts.

ErasStakingPoints
Rewards of stakers for contracts(called by "Dapps Nominator") at era.

ErasTotalStake
The total amount staked for the last HISTORY_DEPTH eras. If total hasn't been set or has been removed then 0 stake is returned.

GenesisConfig
Genesis config for the module, allow to build genesis storage.

Ledger
Map from all (unlocked) "controller" accounts to the info regarding the staking.

Module

Nominations
A record of the nominations made by a specific account.

NominatorsUntreatedEra

OperatorsUntreatedEra

Payee
Where the reward payment should be made. Keyed by stash.

StakingLedger
The ledger of a (bonded) stash.

UnlockChunk
Just a Balance/BlockNumber tuple to encode when a chunk of funds will be unlocked.

UntreatedEra
The already untreated era is EraIndex.

Enums

Call
Dispatchable calls.

Error
Error for the staking module.

Forcing
Mode of era-forcing.

RawEvent
Events for this module.

RewardDestination
A destination account for payment.

Traits

Trait

Type Definitions

BalanceOf

Event
RawEvent specialized for the configuration Trait

MomentOf

SessionIndex
Simple index type with which we can count sessions.