Crate pallet_dapps_staking

Dapps Staking Pallet

Config

Overview

Pallet that implements dapps staking protocol.

Dapps staking protocol is a completely decentralized & innovative approach to reward developers for their contribution to the Astar/Shiden ecosystem. Stakers can pick a dapp and nominate it for rewards by locking their tokens. Dapps will be rewarded, based on the proportion of locked tokens. Stakers are also rewarded, based on the total amount they’ve locked (invariant of the dapp they staked on).

Rewards are accumulated throughout an era and when era finishes, both stakers and developers can claim their rewards for that era. This is a continous process. Rewards can be claimed even for eras which are older than the last one (no limit at the moment).

Reward claiming isn’t automated since the whole process is done on-chain and is fully decentralized. Both stakers and developers are responsible for claiming their own rewards.

Interface

Dispatchable Function

register - used to register a new contract for dapps staking

unregister - used to unregister contract from dapps staking, making it ineligible for receiveing future rewards

withdraw_from_unregistered - used by stakers to withdraw their stake from an unregistered contract (no unbonding period)

bond_and_stake - basic call for nominating a dapp and locking stakers tokens into dapps staking

unbond_and_unstake - removes nomination from the contract, starting the unbonding process for the unstaked funds

withdraw_unbonded - withdraws all funds that have completed the unbonding period

nomination_transfer - transfer nomination from one contract to another contract (avoids unbonding period)

claim_staker - claims staker reward for a single era

claim_dapp - claims dapp rewards for the specified era

force_new_era - forces new era on the start of the next block

maintenance_mode - enables or disables pallet maintenance mode

set_reward_destination - sets reward destination for the staker rewards

set_contract_stake_info - root-only call to set storage value (used for fixing corrupted data)

burn_stale_reward - root-only call to burn unclaimed, stale rewards from unregistered contracts

User is encouraged to refer to specific function implementations for more comprehensive documentation.

Other

on_initialize - part of Hooks trait, it’s important to call this per block since it handles reward snapshots and era advancement.

account_id - returns pallet’s account Id

ensure_pallet_enabled - checks whether pallet is in maintenance mode or not and returns appropriate Result

rewards - used to deposit staker and dapps rewards into dApps staking reward pool

tvl - total value locked in dApps staking (might differ from total staked value)

Re-exports

pub use weights::WeightInfo;

pub use pallet::pallet::*;

Modules

pallet

Dapps staking FRAME Pallet.

weights

Autogenerated weights for pallet_dapps_staking

Structs

AccountLedger

Contains information about account’s locked & unbonding balances.

ContractStakeInfo

Used to split total EraPayout among contracts. Each tuple (contract, era) has this structure. This will be used to reward contracts developer and his stakers.

DAppInfo

EraInfo

A record for total rewards and total amount staked for an era

EraStake

Used to represent how much was staked in a particular era. E.g. {staked: 1000, era: 5} means that in era 5, staked amount was 1000.

RewardInfo

A record of rewards allocated for stakers and dapps

StakerInfo

Used to provide a compact and bounded storage for information about stakes in unclaimed eras.

UnbondingInfo

Contains unlocking chunks. This is a convenience struct that provides various utility methods to help with unbonding handling.

UnlockingChunk

Represents an balance amount undergoing the unbonding process. Since unbonding takes time, it’s important to keep track of when and how much was unbonded.

Enums

DAppState

DApp State descriptor

Forcing

Mode of era-forcing.

RewardDestination

Instruction on how to handle reward payout for stakers. In order to make staking more competitive, majority of stakers will want to automatically restake anything they earn.

Version

Storage value representing the current Dapps staking pallet storage version. Used by on_runtime_upgrade to determine whether a storage migration is needed or not.

Type Definitions

EraIndex

Counter for the number of eras that have passed.