Module pallet_dapps_staking::pallet::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

bond_and_stake

Lock up and stake balance of the origin account.

burn_stale_reward

Used to burn unclaimed & stale rewards from an unregistered contract.

claim_dapp

Claim earned dapp rewards for the specified era.

claim_staker

Claim earned staker rewards for the oldest unclaimed era. In order to claim multiple eras, this call has to be called multiple times.

claim_staker_for

Claim earned staker rewards for the given staker, and the oldest unclaimed era. In order to claim multiple eras, this call has to be called multiple times.

decommission

Enable the decommission flag for the pallet.

force_new_era

Force a new era at the start of the next block.

maintenance_mode

true will disable pallet, enabling maintenance mode. false will do the opposite.

nomination_transfer

Transfer nomination from one contract to another.

register

Used to register contract for dapps staking. The origin account used is treated as the developer account.

set_contract_stake_info

Used to force set ContractEraStake storage values. The purpose of this call is only for fixing one of the issues detected with dapps-staking.

set_reward_destination

Used to set reward destination for staker rewards.

set_reward_destination_for

Used to set reward destination for staker rewards, for the given staker

unbond_and_unstake

Start unbonding process and unstake balance from the contract.

unregister

Unregister existing contract from dapps staking, making it ineligible for rewards from current era onwards. This must be called by the root (at the moment).

withdraw_from_unregistered

Withdraw locked funds from a contract that was unregistered.

withdraw_unbonded

Withdraw all funds that have completed the unbonding process.