Trait pallet_dapps_staking::pallet::pallet::Config

pub trait Config: Config {

    type Currency: Currency<Self::AccountId, Balance = Balance> + LockableCurrency<Self::AccountId, Moment = Self::BlockNumber> + ReservableCurrency<Self::AccountId>;
    type SmartContract: Default + Parameter + Member + MaxEncodedLen;
    type BlockPerEra: Get<BlockNumberFor<Self>>;
    type RegisterDeposit: Get<Balance>;
    type MaxNumberOfStakersPerContract: Get<u32>;
    type MinimumStakingAmount: Get<Balance>;
    type PalletId: Get<PalletId>;
    type MinimumRemainingAmount: Get<Balance>;
    type MaxUnlockingChunks: Get<u32>;
    type UnbondingPeriod: Get<u32>;
    type MaxEraStakeValues: Get<u32>;
    type UnregisteredDappRewardRetention: Get<u32>;
    type ForcePalletDisabled: Get<bool>;
    type DelegateClaimFee: Get<Balance>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Currency: Currency<Self::AccountId, Balance = Balance> + LockableCurrency<Self::AccountId, Moment = Self::BlockNumber> + ReservableCurrency<Self::AccountId>

The staking balance.

type SmartContract: Default + Parameter + Member + MaxEncodedLen

Describes smart contract in the context required by dapps staking.

type BlockPerEra: Get<BlockNumberFor<Self>>

Number of blocks per era.

type RegisterDeposit: Get<Balance>

Deposit that will be reserved as part of new contract registration.

type MaxNumberOfStakersPerContract: Get<u32>

Maximum number of unique stakers per contract.

type MinimumStakingAmount: Get<Balance>

Minimum amount user must have staked on contract. User can stake less if they already have the minimum staking amount staked on that particular contract.

type PalletId: Get<PalletId>

Dapps staking pallet Id

type MinimumRemainingAmount: Get<Balance>

Minimum amount that should be left on staker account after staking. Serves as a safeguard to prevent users from locking their entire free balance.

type MaxUnlockingChunks: Get<u32>

Max number of unlocking chunks per account Id <-> contract Id pairing. If value is zero, unlocking becomes impossible.

type UnbondingPeriod: Get<u32>

Number of eras that need to pass until unstaked value can be withdrawn. Current era is always counted as full era (regardless how much blocks are remaining). When set to 0, it’s equal to having no unbonding period.

type MaxEraStakeValues: Get<u32>

Max number of unique EraStake values that can exist for a (staker, contract) pairing. When stakers claims rewards, they will either keep the number of EraStake values the same or they will reduce them by one. Stakers cannot add an additional EraStake value by calling bond&stake or unbond&unstake if they’ve reached the max number of values.

This ensures that history doesn’t grow indefinitely - if there are too many chunks, stakers should first claim their former rewards before adding additional EraStake values.

type UnregisteredDappRewardRetention: Get<u32>

Number of eras that need to pass until dApp rewards for the unregistered contracts can be burned. Developer can still claim rewards after this period has passed, iff it hasn’t been burned yet.

For example, if retention is set to 2 and current era is 10, it means that all unclaimed rewards bellow era 8 can be burned.

type ForcePalletDisabled: Get<bool>

Can be used to force pallet into permanent maintenance mode.

type DelegateClaimFee: Get<Balance>

The fee that will be charged for claiming rewards on behalf of a staker. This amount will be transferred from the staker over to the caller.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.